Version 1.0 Final
This commit marks the final version 1.0 of the project. The code included in this commit has been developed as part of my project work at Siemens Energy in collaboration with DHBW Karlsruhe. All features and functionalities planned for this version have been implemented and thoroughly tested. This version is now ready for deployment and further use.